test srv was sending 64, client was only reading a 32, on 64 bit platform -- modified the client to read a size_t like it should

+++ src/test_client.cpp
@@ -25,7 +25,7 @@ static int send( int sock, int nb )
 
 static int recv( int sock )
 {
-   uint32_t msg = 0;
+   size_t msg = 0;
    SysErrIf( recv( sock, &msg, sizeof(msg), 0 ) == -1 );
    msg = ntohl(msg); 
    std::cout << "recv: " << msg << "\n";
